# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5dcc4fd7f1ef4518a5383351
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcc504ff1ef4518a538335b
  title: OrderManagment
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Order  -----
  /order/:
    get:
      x-skaffolder-id: 5dcc52a6f1ef4518a53833db
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcc5287f1ef4518a53833d2
      x-skaffolder-resource: Order
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Order
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    
  #-----  Product  -----
  /product/:
    post:
      x-skaffolder-id: 5dcc52a6f1ef4518a53833d6
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Product
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Price:
                  type: integer
                Type:
                  type: string
    get:
      x-skaffolder-id: 5dcc52a6f1ef4518a53833de
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Product
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    
  /product/{id}:
    delete:
      x-skaffolder-id: 5dcc52a6f1ef4518a53833dd
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Product
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcc52a6f1ef4518a53833d8
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Product
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    post:
      x-skaffolder-id: 5dcc52a6f1ef4518a53833d7
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-resource: Product
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Product
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:

  # Security
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
        

    Order:
      x-skaffolder-id: 5dcc5287f1ef4518a53833d2
      x-skaffolder-id-db: 5dcc504ff1ef4518a538335c
      x-skaffolder-id-entity: 5dcc5287f1ef4518a53833cc
      x-skaffolder-url: /order
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        amount:
          type: integer
          x-skaffolder-id-attr: 5dcc5287f1ef4518a53833d0
          x-skaffolder-type: Integer
          x-skaffolder-required: true
        customerID:
          type: integer
          x-skaffolder-id-attr: 5dcc5287f1ef4518a53833cd
          x-skaffolder-type: Integer
          x-skaffolder-required: true
        customerName:
          type: string
          x-skaffolder-id-attr: 5dcc5287f1ef4518a53833ce
          x-skaffolder-type: String
          x-skaffolder-required: true
        orderDate:
          type: integer
          x-skaffolder-id-attr: 5dcc5287f1ef4518a53833d1
          x-skaffolder-type: Date
          x-skaffolder-required: true
        productID:
          type: integer
          x-skaffolder-id-attr: 5dcc5287f1ef4518a53833cf
          x-skaffolder-type: Integer
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - amount
        - customerID
        - customerName
        - orderDate
        - productID
      
    
    Product:
      x-skaffolder-id: 5dcc5105f1ef4518a53833c7
      x-skaffolder-id-db: 5dcc504ff1ef4518a538335c
      x-skaffolder-id-entity: 5dcc5105f1ef4518a53833c3
      x-skaffolder-url: /product
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Price:
          type: integer
          x-skaffolder-id-attr: 5dcc5105f1ef4518a53833c5
          x-skaffolder-type: Integer
          x-skaffolder-required: true
        Type:
          type: string
          x-skaffolder-id-attr: 5dcc5105f1ef4518a53833c4
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
        has:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dcc5105f1ef4518a53833c3
          x-skaffolder-ent2: 5dcc5287f1ef4518a53833cc
      required:
        - _id
        - Price
        - Type
      
    
    
  # Databases
  x-skaffolder-db:
    - x-skaffolder-id: 5dcc504ff1ef4518a538335c
      x-skaffolder-name: OrderManagment_db

  # Pages
  x-skaffolder-page:
    - x-skaffolder-id: 5dcc504ff1ef4518a5383364
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5dcc52a6f1ef4518a53833dc
      x-skaffolder-roles: []
    - x-skaffolder-id: 5dcc52a6f1ef4518a53833d5
      x-skaffolder-name: ProductEdit
      x-skaffolder-url: /products/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-services: 
        - 5dcc52a6f1ef4518a53833d6
        - 5dcc52a6f1ef4518a53833d8
        - 5dcc52a6f1ef4518a53833db
        - 5dcc52a6f1ef4518a53833d7
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5dcc52a6f1ef4518a53833dc
      x-skaffolder-name: ProductList
      x-skaffolder-url: /products
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5dcc5105f1ef4518a53833c7
      x-skaffolder-services: 
        - 5dcc52a6f1ef4518a53833dd
        - 5dcc52a6f1ef4518a53833de
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5dcc52a6f1ef4518a53833d5
      x-skaffolder-roles: []